requirements.txt
tensorflow-gpu
keras
h5py
scipy
pillow
imageio
